backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.29.7
django-storages==1.14.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
